+++ controllers/front/return.php
@@ -230,7 +230,7 @@ class MollieReturnModuleFrontController extends AbstractMollieController
         switch ($orderStatus) {
             case PaymentStatus::STATUS_OPEN:
             case PaymentStatus::STATUS_PENDING:
-                if ($transaction->mode === "test") {
+                if ($transaction->mode === 'test') {
                     $this->setWarning($this->module->l('We have not received a definite payment status. You will be notified as soon as we receive a confirmation of the bank/merchant.'));
                     $response = $paymentReturnService->handleTestPendingStatus();
                     break;