+++ upgrade/Upgrade-3.3.2.php
@@ -8,6 +8,7 @@
  *
  * @see        https://github.com/mollie/PrestaShop
  */
+
 if (!defined('_PS_VERSION_')) {
     exit;
 }